requirements.txt
matplotlib===3.5.2
torch===1.12.0
torchvision===0.13.0
vit-pytorch===0.35.8
pandas===1.4.3
Pillow===9.2.0